+++ src/fitter/GibbsSamplerFitter.py
@@ -1,0 +1,107 @@
+from PIL import Image
+from numpy import ones, zeros, linspace, argmin, nonzero, column_stack, exp, array
+from numpy.random import rand, randint
+
+from .ModelFitter import ModelFitter
+
+class GibbsSamplerFitter(ModelFitter):
+    def __init__(self, image, dimensions=199, model=None, steps=None,
+                 max_loops=1, determined_loops=0):
+        self.__steps = ones(dimensions, dtype='i') if steps is None else steps
+        self.__current_step = None
+        self.__parameters = zeros(dimensions, dtype='f')
+
+        self.__errors = None
+        self.__values = None
+        self.__max_loops = max_loops
+        self.__determined_loops = determined_loops
+
+        super(GibbsSamplerFitter, self).__init__(image, dimensions, model)
+
+    def start(self):
+        self.__loop = 0
+        self.request_normals(zeros(self._dimensions, dtype='f'), 'init')
+        self.__get_parameter(0)
+
+    def __get_parameter(self, i):
+        self.__current_step = i
+
+        if self.__steps[i] % 2 == 0:
+            self.__values = linspace(-4, 4, self.__steps[i] + 1, dtype='f')
+        else:
+            self.__values = linspace(-4, 4, self.__steps[i] + 2, dtype='f')
+
+        self.__errors = [None] * self.__values.size
+        for i, value in enumerate(self.__values):
+            self.__values[i] = value
+            # parameters = self.__parameters.copy()
+            # parameters[self.__current_step] = value
+            # self.request_normals(parameters, i)
+            self.request_normals((self.__current_step, value), i)
+
+    def receive_normals(self, normals, index=None):
+        if index == 'init':
+            return
+        elif index == 'pre':
+            self.__get_parameter(self.__current_step + 1)
+            return
+
+        light = self.estimate_light(normals)
+        shadows = normals.dot(light)
+
+        if index is None:
+            self.finish(normals, shadows)
+            return
+
+        self.__errors[index] = self.get_image_deviation(shadows, normals)
+
+        if None in self.__errors:
+            return
+
+        errors = array(self.__errors)
+        max_error = errors.max()
+        if self.__loop >= self.__determined_loops:
+            X = exp(- errors / max_error).sum()
+            v = rand()
+            best_index = -1
+            t = 0
+            for i, error in enumerate(errors):
+                e = exp(error / max_error) / X
+                t += e
+                if v <= e:
+                    best_index = i
+                    break
+                else:
+                    v -= e
+            if best_index == -1:
+                best_index = len(self.__errors) - 1
+        else:
+            best_index = argmin(self.__errors)
+
+        self.__parameters[self.__current_step] = self.__values[best_index]
+        print('{}.{}: Error of the best is {} ({})'.format(
+            self.__loop, self.__current_step, self.__errors[best_index],
+            min(self.__errors)))
+
+        if self.__current_step + 1 == self._dimensions and self.__loop == 2:
+            self.request_normals(self.__parameters)
+            return
+        elif self.__current_step + 1 == self._dimensions:
+            self.__current_step = 0
+            self.__loop += 1
+
+        self.request_normals((self.__current_step, self.__values[best_index]),
+                             'pre')
+
+    def finish(self, normals, shadows):
+        img = shadows[::-1]
+        alpha = zeros(len(normals[:, 3]))
+        alpha[nonzero(normals[:, 3])[0]] = 1.
+        data = (column_stack((
+                img, img, img,
+                alpha
+                )) * 255).astype('i')
+        image = Image.new('L', (500, 500))
+        image.putdata((img*255).astype('i'))
+        image.save('img.png')
+        print('Finished')